refactor: common to utils on error
