+++ src/common/error/error-classes.ts
