+++ src/common/error/error-wrapper.test.ts
@@ -1,8 +1,8 @@
 import { AuthError } from "next-auth";
 import { describe, expect, test, vi } from "vitest";
+import { PushoverError, UnexpectedError } from "@/common/error/error-classes";
 import { Prisma } from "@/generated";
 import { serverLogger } from "@/o11y/server";
-import { PushoverError, UnexpectedError } from "@/utils/error/error-classes";
 import { wrapServerSideErrorForClient } from "./error-wrapper";
 
 describe("wrapServerSideErrorForClient", () => {
