+++ src/features/books/actions/add-books.ts
@@ -2,11 +2,11 @@
 import "server-only";
 import { revalidatePath } from "next/cache";
 import { forbidden } from "next/navigation";
+import { wrapServerSideErrorForClient } from "@/common/error/error-wrapper";
 import { BooksDomainService } from "@/domains/books/services/books-domain-service";
 import { booksCommandRepository } from "@/infrastructures/books/repositories/books-command-repository";
 import { booksQueryRepository } from "@/infrastructures/books/repositories/books-query-repository";
 import { getSelfId, hasDumperPostPermission } from "@/utils/auth/session";
-import { wrapServerSideErrorForClient } from "@/utils/error/error-wrapper";
 import type { ServerAction } from "@/utils/types";
 
 export async function addBooks(formData: FormData): Promise<ServerAction> {
