+++ src/features/news/actions/add-news.ts
@@ -2,11 +2,11 @@
 import "server-only";
 import { revalidatePath } from "next/cache";
 import { forbidden } from "next/navigation";
+import { wrapServerSideErrorForClient } from "@/common/error/error-wrapper";
 import { NewsDomainService } from "@/domains/news/services/news-domain-service";
 import { newsCommandRepository } from "@/infrastructures/news/repositories/news-command-repository";
 import { newsQueryRepository } from "@/infrastructures/news/repositories/news-query-repository";
 import { getSelfId, hasDumperPostPermission } from "@/utils/auth/session";
-import { wrapServerSideErrorForClient } from "@/utils/error/error-wrapper";
 import type { ServerAction } from "@/utils/types";
 
 export async function addNews(formData: FormData): Promise<ServerAction> {
