+++ src/features/news/actions/delete-news.ts
@@ -2,9 +2,9 @@
 import "server-only";
 import { revalidatePath } from "next/cache";
 import { forbidden } from "next/navigation";
+import { wrapServerSideErrorForClient } from "@/common/error/error-wrapper";
 import { newsCommandRepository } from "@/infrastructures/news/repositories/news-command-repository";
 import { getSelfId, hasDumperPostPermission } from "@/utils/auth/session";
-import { wrapServerSideErrorForClient } from "@/utils/error/error-wrapper";
 import type { ServerAction } from "@/utils/types";
 
 export async function deleteNews(id: string): Promise<ServerAction> {
